moved HandCubeTab into separate file
