+++ ExternalPCConnection/gloveController.py
@@ -18,205 +18,10 @@ from Utility.SerialThread import SerialThread
 import Utility.SerialThread # to access constants
 from Utility.LoggerThread import LoggerThread
 from classification.GraspInferenceThread import GraspInferenceThread
+from PyQtTabs.HandCubeTab import HandCubeTab
 
 COM_PORT = 'COM16'  # <-- CHANGE THIS TO YOUR TEENSY PORT
 BAUD_RATE = 115200
-
-# --- 3D Plotting ---
-class HandCubeTab(QWidget):
-    """3D viewer for hand landmarks and the tracked cube, with visibility toggles."""
-
-    # Standard 21-point hand landmark indices (per LANDMARK_NAMES in LeapConnector.py)
-    FINGERTIP_INDICES = [4, 8, 12, 16, 20]  # thumb, index, middle, ring, pinky tips
-    CUBE_HALF_EXTENT = 0.02  # meters
-
-    def __init__(self, data_manager, thresholds, parent=None):
-        super().__init__(parent)
-        self.dm = data_manager
-        self.thresholds = thresholds
-        self.show_fingertips_only = False
-        self.show_cube = True
-        self.show_hand = True
-        self.latest_tripod = None # finger tips
-        self.latest_cube_pos = None
-        self._build_ui()
-
-    def _build_ui(self):
-        layout = QVBoxLayout(self)
-
-        controls = QHBoxLayout()
-        self.chk_show_hand = QCheckBox("Show Hand")
-        self.chk_show_hand.setChecked(True)
-        self.chk_show_hand.stateChanged.connect(self._on_hand_toggle)
-        controls.addWidget(self.chk_show_hand)
-
-        self.chk_fingertips_only = QCheckBox("Fingertips Only")
-        self.chk_fingertips_only.stateChanged.connect(self._on_fingertips_toggle)
-        controls.addWidget(self.chk_fingertips_only)
-
-        self.chk_show_cube = QCheckBox("Show Cube")
-        self.chk_show_cube.setChecked(True)
-        self.chk_show_cube.stateChanged.connect(self._on_cube_toggle)
-        controls.addWidget(self.chk_show_cube)
-
-        controls.addStretch()
-        layout.addLayout(controls)
-
-        self.view = gl.GLViewWidget()
-        self.view.setCameraPosition(distance=0.5)
-        grid = gl.GLGridItem()
-        grid.setSize(1, 1)
-        grid.setSpacing(0.05, 0.05)
-        self.view.addItem(grid)
-        layout.addWidget(self.view)
-
-        self.hand_scatter = gl.GLScatterPlotItem(pos=np.zeros((1, 3)), color=(0.2, 0.8, 1.0, 1.0), size=8)
-        self.view.addItem(self.hand_scatter)
-
-        # 12 edges of the cube, drawn as a wireframe via GLLinePlotItem(mode='lines')
-        self.cube_lines = gl.GLLinePlotItem(pos=np.zeros((2, 3)), color=(1.0, 0.4, 0.1, 1.0), width=2, mode='lines')
-        self.view.addItem(self.cube_lines)
-
-        thresh_group = QGroupBox("Proximity Thresholds")
-        thresh_layout = QGridLayout()
-
-        thresh_layout.addWidget(QLabel("Cube Proximity Threshold:"), 0, 0)
-        self.sld_cube_thresh = QSlider(Qt.Orientation.Horizontal)
-        self.sld_cube_thresh.setRange(1, 500)  # millimeters
-        self.sld_cube_thresh.setValue(int(self.thresholds.cube_proximity * 1000))
-        self.sld_cube_thresh.valueChanged.connect(self._on_cube_thresh_slider)
-        thresh_layout.addWidget(self.sld_cube_thresh, 0, 1)
-        self.lbl_cube_thresh_val = QLabel(f"{self.thresholds.cube_proximity:.3f} m")
-        thresh_layout.addWidget(self.lbl_cube_thresh_val, 0, 2)
-        thresh_layout.addWidget(QLabel("Min Tip-to-Cube Dist:"), 0, 3)
-        self.lbl_cube_dist = QLabel("--")
-        thresh_layout.addWidget(self.lbl_cube_dist, 0, 4)
-
-        thresh_layout.addWidget(QLabel("Finger Proximity Threshold:"), 1, 0)
-        self.sld_finger_thresh = QSlider(Qt.Orientation.Horizontal)
-        self.sld_finger_thresh.setRange(1, 500)  # millimeters
-        self.sld_finger_thresh.setValue(int(self.thresholds.finger_proximity * 1000))
-        self.sld_finger_thresh.valueChanged.connect(self._on_finger_thresh_slider)
-        thresh_layout.addWidget(self.sld_finger_thresh, 1, 1)
-        self.lbl_finger_thresh_val = QLabel(f"{self.thresholds.finger_proximity:.3f} m")
-        thresh_layout.addWidget(self.lbl_finger_thresh_val, 1, 2)
-        thresh_layout.addWidget(QLabel("Max Finger-to-Finger Dist:"), 1, 3)
-        self.lbl_finger_dist = QLabel("--")
-        thresh_layout.addWidget(self.lbl_finger_dist, 1, 4)
-
-        thresh_group.setLayout(thresh_layout)
-        layout.addWidget(thresh_group)
-
-    def _on_cube_thresh_slider(self, value_mm):
-        self.thresholds.cube_proximity = value_mm / 1000.0
-        self.lbl_cube_thresh_val.setText(f"{self.thresholds.cube_proximity:.3f} m")
-
-    def _on_finger_thresh_slider(self, value_mm):
-        self.thresholds.finger_proximity = value_mm / 1000.0
-        self.lbl_finger_thresh_val.setText(f"{self.thresholds.finger_proximity:.3f} m")
-
-    def _on_hand_toggle(self, state):
-        self.show_hand = bool(state)
-        self.hand_scatter.setVisible(self.show_hand)
-
-    def _on_fingertips_toggle(self, state):
-        self.show_fingertips_only = bool(state)
-
-    def _on_cube_toggle(self, state):
-        self.show_cube = bool(state)
-        self.cube_lines.setVisible(self.show_cube)
-
-    def _update_threshold_readout(self):
-        if self.latest_tripod is None:
-            return
-
-        thumb = np.array(self.latest_tripod['thumb'])
-        index = np.array(self.latest_tripod['index'])
-        middle = np.array(self.latest_tripod['middle'])
-
-        finger_dist = max(
-            np.linalg.norm(thumb - index),
-            np.linalg.norm(thumb - middle),
-            np.linalg.norm(index - middle),
-        )
-        finger_ok = finger_dist < self.thresholds.finger_proximity
-        self.lbl_finger_dist.setText(f"{finger_dist:.3f} m")
-        self.lbl_finger_dist.setStyleSheet(f"color: {'#2ecc71' if finger_ok else '#e74c3c'}; font-weight: bold;")
-
-        if self.latest_cube_pos is not None:
-            cube_dist = max(
-                np.linalg.norm(thumb - self.latest_cube_pos),
-                np.linalg.norm(index - self.latest_cube_pos),
-                np.linalg.norm(middle - self.latest_cube_pos),
-            )
-            cube_ok = cube_dist < self.thresholds.cube_proximity
-            self.lbl_cube_dist.setText(f"{cube_dist:.3f} m")
-            self.lbl_cube_dist.setStyleSheet(f"color: {'#2ecc71' if cube_ok else '#e74c3c'}; font-weight: bold;")
-
-    @staticmethod
-    def _cube_corners(tx, ty, tz, rx, ry, rz, half_extent):
-        cx, cy, cz = np.cos([rx, ry, rz])
-        sx, sy, sz = np.sin([rx, ry, rz])
-        Rx = np.array([[1, 0, 0], [0, cx, -sx], [0, sx, cx]])
-        Ry = np.array([[cy, 0, sy], [0, 1, 0], [-sy, 0, cy]])
-        Rz = np.array([[cz, -sz, 0], [sz, cz, 0], [0, 0, 1]])
-        R = Rz @ Ry @ Rx
-
-        h = half_extent
-        local_corners = np.array([
-            [-h, -h, -h], [h, -h, -h], [h, h, -h], [-h, h, -h],
-            [-h, -h, h], [h, -h, h], [h, h, h], [-h, h, h],
-        ])
-        return (R @ local_corners.T).T + np.array([tx, ty, tz])
-
-    @staticmethod
-    def _cube_edges(corners):
-        edge_idx = [
-            (0, 1), (1, 2), (2, 3), (3, 0),  # bottom face
-            (4, 5), (5, 6), (6, 7), (7, 4),  # top face
-            (0, 4), (1, 5), (2, 6), (3, 7),  # verticals
-        ]
-        pts = []
-        for i, j in edge_idx:
-            pts.append(corners[i])
-            pts.append(corners[j])
-        return np.array(pts)
-
-    def update_view(self):
-        # --- Hand points ---
-        hp_queue = self.dm.subscribers['gui_hp']
-        latest_hands = None
-        while not hp_queue.empty():
-            _, hands = hp_queue.get()
-            latest_hands = hands
-
-        if latest_hands is not None and self.show_hand:
-            pts = []
-            for hand in latest_hands:
-                kp = hand.get("keypoints_list")
-                if not kp:
-                    continue
-                indices = self.FINGERTIP_INDICES if self.show_fingertips_only else range(len(kp))
-                for idx in indices:
-                    if idx < len(kp):
-                        pts.append(kp[idx])
-            if pts:
-                self.hand_scatter.setData(pos=np.array(pts))
-
-        # --- Cube ---
-        cube_queue = self.dm.subscribers['gui_cube']
-        latest_cube = None
-        while not cube_queue.empty():
-            _, cube = cube_queue.get()  # (timestamp, [tx,ty,tz,rx,ry,rz])
-            latest_cube = cube
-
-        if latest_cube is not None:
-            self.latest_cube_pos = np.array(latest_cube[:3])
-            if self.show_cube:
-                corners = self._cube_corners(*latest_cube, self.CUBE_HALF_EXTENT) # type: ignore
-                self.cube_lines.setData(pos=self._cube_edges(corners))
-
-        self._update_threshold_readout()
 
 # --- Main PyQt6 GUI ---
 class MainWindow(QMainWindow):
